+++ src/Aardvark.Base.FSharp/Utilities/Interop/FSLibExtensions.fs
@@ -292,6 +292,12 @@ module Prelude =
                 NativePtr.set ptr i arr.[i]
             ptr
 
+        /// Allocates and initializes native memory on the stack from the given array and mapping function.
+        let inline stackUseArr ([<InlineIfLambda>] mapping: 'T -> 'U) (data: 'T[]) =
+            let ptr = NativePtr.stackalloc<'U> data.Length
+            for i = 0 to data.Length - 1 do ptr.[i] <- mapping data.[i]
+            ptr
+
         let toStream (count : int) (ptr : nativeptr<'a>) : Stream =
             let l = int64 <| sizeof<'a> * count
             new System.IO.UnmanagedMemoryStream(cast ptr, l,l, FileAccess.ReadWrite) :> _