+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 - [FSharp] Added assignment operators
 - [FSharp] Added Type.CLRSize extension property
 - [Introspection] Made native library loading more robust
+- [Introspection] Fixed issue with plugin cache invalidation
 
 ### 5.3.8
 - [FSharp] Added stableSum and stableSumBy for List, Array, and Seq